requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
pydantic